review comments

+++ samples/go/photo-date-dedup/main.go
@@ -161,7 +161,7 @@ func buildGroups(db types.ValueReadWriter, thresh int, byDate types.Map) types.L
 	byDate.IterAll(func(key, s types.Value) {
 		s.(types.Set).IterAll(func(val types.Value) {
 			dt := float64(key.(types.Number))
-			if dt == float64(0) {
+			if dt == 0 {
 				// If date is not known, then the photo is in its own group
 				flush()
 				startGroup(val)
